% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{data}
\name{fit_imputed}
\alias{fit_imputed}
\title{Fitted Ordered Beta Regression Model (Imputed Datasets)}
\format{
an \code{ordbetareg} object
}
\usage{
fit_imputed
}
\description{
A fitted ordered beta regression model on multiple imputed
datasets generated by the package \code{mice}.
}
\keyword{datasets}
