% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{data}
\name{fit_multivariate}
\alias{fit_multivariate}
\title{Fitted Ordered Beta Regression Model (Multivariate regression)}
\format{
an \code{ordbetareg} object
}
\usage{
fit_multivariate
}
\description{
A fitted ordered beta regression model with two responses,
one an ordered beta regression and the other a Gaussian/Normal
outcome. Useful for examining mediation analysis.
}
\keyword{datasets}
