% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{data}
\name{ord_fit_mean}
\alias{ord_fit_mean}
\title{Fitted Ordered Beta Regression Model}
\format{
an \code{ordbetareg} object
}
\usage{
ord_fit_mean
}
\description{
A fitted ordered beta regression model to the mean of the thermometer
column from the pew data.
}
\keyword{datasets}
