% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{data}
\name{ord_fit_phi}
\alias{ord_fit_phi}
\title{Fitted Ordered Beta Regression Model (Phi Regression)}
\format{
an \code{ordbetareg} object
}
\usage{
ord_fit_phi
}
\description{
A fitted ordered beta regression model to the dispersion parameter
of the thermometer
column from the pew data.
}
\keyword{datasets}
