% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{package}
\name{ordbetareg-package}
\alias{ordbetareg-package}
\title{ordbetareg: A Model for Continuous Data with Lower and Upper Bounds}
\description{
The \code{ordbetareg} package is essentially a wrapper around \code{brms} that
enables the ordered beta regression model to be fit. This model has
advantages over other alternatives for continous data with upper
and lower bounds, such as survey sliders, indexes,
dose-response relationships,
and visual analog scales (among others). The package allows for all of the
many \code{brms} regression modeling functions to be used with the ordered
beta regression distribution.
}
\details{
To learn more about how the package works, see the vignette by using
the command \code{browseVignettes(package='ordbetareg')}.

For more info about the distribution, see
this paper: https://osf.io/preprints/socarxiv/2sx6y/

To cite the package, please cite the following paper:

Kubinec, Robert. "Ordered Beta Regression: A Parsimonious, Well-Fitting Model for Continuous Data with Lower and Upper Bounds." \strong{Political Analysis}. 2022.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/saudiwin/ordbetareg_pack/issues}
}

}
\author{
\strong{Maintainer}: Robert Kubinec \email{bobkubinec@gmail.com} (\href{https://orcid.org/0000-0001-6655-4119}{ORCID})

}
\keyword{internal}
