% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{data}
\name{pew}
\alias{pew}
\title{Pew American Trends Panel Wave 28}
\format{
A data frame with 140 variables and 2,538 observations.
}
\source{
https://www.pewresearch.org/social-trends/dataset/american-trends-panel-wave-28/]
}
\usage{
pew
}
\description{
A dataset with the non-missing responses for the 28th wave of the
Pew American Trends Panel survey.
}
\keyword{datasets}
