% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Simulated Ordered Beta Regression Values}
\format{
A dataframe
}
\usage{
sim_data
}
\description{
The simulated draws used in the vignette for calculating statistical
power.
}
\keyword{datasets}
