% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{dordbeta}
\alias{dordbeta}
\title{Probability Density Function for the Ordered Beta Distribution}
\usage{
dordbeta(x = 0.9, mu = 0.5, phi = 1, cutpoints = c(-1, 1), log = FALSE)
}
\arguments{
\item{x}{Variates of the ordered beta distribution (should be in the [0,1] interval).}

\item{mu}{Value of the mean of the distribution.
Should be in the \(0,1\) interval (cannot be strictly equal to 0 or 1). If
length is greater than 1, should be of length x.}

\item{phi}{Value of the dispersion parameter. Should be strictly greater than 0. If
length is greater than 1, should be of length x.}

\item{cutpoints}{A vector of two numeric values for the cutpoints. Second value should}

\item{log}{where to return the log density
be strictly greater than the first value.}
}
\description{
This function will return the density of given variates of the
ordered beta distribution conditional on values for
the mean (\code{mu}), dispersion (\code{phi}) and cutpoints
governing the ratio of degenerate (discrete) to continuous
responses.
}
